STRY0010006
The Loaner Vehicle Catalog table needs to be filled in with a list of the current
vehicle inventory of the company. Import a spreadsheet into the Loaner Vehicle Catalog table and map the fields accordingly.

I completed this user story by navigating to All>Loaner Vehicle Request and Inventory>Loaner Vehicle Catalog>All, right-clicking in the column header, and selecting Import.