STRY0010007
Every time a vehicle is requested to be used by a contractor, we would need to
track who requested it and what office is associated with the person who requested for it.
This will be our method of tracking the whereabouts of the vehicle. For us to do this we will need a table that would store the important information we would need for tracking.