STRY0010013
We would like to have buttons for users to use to decide on an action when the Vehicles have been returned. Users should be able to click a Return to Warehouse
button when the user sees that the vehicle is fit to be returned to the warehouse. We should also have a button called Send to Repair if the vehicle needs to be repaired or serviced before returning to the Warehouse.